requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
djangorestframework==3.14.0
pycodestyle==2.10.0
pytube==12.1.3
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
